Chore[COMP-78]: Refactor decorators
