+++ ProjectSourceCode/Decorator/abstractcomponent.h
