+++ ProjectSourceCode/Decorator/componentdecorator.cpp
