+++ ProjectSourceCode/Decorator/componentdecorator.h
